modules structure draft
