+++ rectools/models/nn/bert4rec.py
@@ -5,28 +5,28 @@ import numpy as np
 import torch
 from pytorch_lightning import Trainer
 from pytorch_lightning.accelerators import Accelerator
-from torch import nn
 
-from rectools.models.sasrec import (
-    CatFeaturesItemNet,
-    IdEmbeddingsItemNet,
-    ItemNetBase,
-    LearnableInversePositionalEncoding,
-    PointWiseFeedForward,
-    PositionalEncodingBase,
-    SessionEncoderDataPreparatorBase,
+from rectools.models.nn.transformer_base import (
+    PADDING_VALUE,
     SessionEncoderLightningModule,
     SessionEncoderLightningModuleBase,
-    TransformerLayersBase,
     TransformerModelBase,
 )
 
-PADDING_VALUE = "PAD"
+from .item_net import CatFeaturesItemNet, IdEmbeddingsItemNet, ItemNetBase
+from .net_blocks import (
+    LearnableInversePositionalEncoding,
+    PositionalEncodingBase,
+    PreLNTransformerLayers,
+    TransformerLayersBase,
+)
+from .transformer_data_preparator import SessionEncoderDataPreparatorBase
+
 MASKING_VALUE = "MASK"
 
 
 class BERT4RecDataPreparator(SessionEncoderDataPreparatorBase):
-    """TODO"""
+    """Data Preparator for BERT4RecModel."""
 
     def __init__(
         self,
@@ -101,56 +101,78 @@ class BERT4RecDataPreparator(SessionEncoderDataPreparatorBase):
         return {"x": torch.LongTensor(x)}
 
 
-class BERT4RecTransformerLayers(TransformerLayersBase):
-    """TODO"""
-
-    def __init__(
-        self,
-        n_blocks: int,
-        n_factors: int,
-        n_heads: int,
-        dropout_rate: float,
-    ):
-        super().__init__()
-        self.n_blocks = n_blocks
-        self.multi_head_attn = nn.ModuleList(
-            [nn.MultiheadAttention(n_factors, n_heads, dropout_rate, batch_first=True) for _ in range(n_blocks)]
-        )
-        self.layer_norm1 = nn.ModuleList([nn.LayerNorm(n_factors) for _ in range(n_blocks)])
-        self.dropout1 = nn.ModuleList([nn.Dropout(dropout_rate) for _ in range(n_blocks)])
-        self.layer_norm2 = nn.ModuleList([nn.LayerNorm(n_factors) for _ in range(n_blocks)])
-        self.feed_forward = nn.ModuleList(
-            [PointWiseFeedForward(n_factors, n_factors * 4, dropout_rate, torch.nn.GELU()) for _ in range(n_blocks)]
-        )
-        self.dropout2 = nn.ModuleList([nn.Dropout(dropout_rate) for _ in range(n_blocks)])
-        self.dropout3 = nn.ModuleList([nn.Dropout(dropout_rate) for _ in range(n_blocks)])
-
-    def forward(
-        self, seqs: torch.Tensor, timeline_mask: torch.Tensor, attn_mask: torch.Tensor, key_padding_mask: torch.Tensor
-    ) -> torch.Tensor:
-        """TODO"""
-        for i in range(self.n_blocks):
-            mha_input = self.layer_norm1[i](seqs)
-            mha_output, _ = self.multi_head_attn[i](
-                mha_input,
-                mha_input,
-                mha_input,
-                attn_mask=attn_mask,
-                key_padding_mask=key_padding_mask,
-                need_weights=False,
-            )
-            seqs = seqs + self.dropout1[i](mha_output)
-            ff_input = self.layer_norm2[i](seqs)
-            ff_output = self.feed_forward[i](ff_input)
-            seqs = seqs + self.dropout2[i](ff_output)
-            seqs = self.dropout3[i](seqs)
-        # TODO: test with torch.nn.Linear and cross-entropy loss as in
-        # https://github.com/jaywonchung/BERT4Rec-VAE-Pytorch/blob/f66f2534ebfd937778c7174b5f9f216efdebe5de/models/bert.py#L11C1-L11C67
-        return seqs
-
-
 class BERT4RecModel(TransformerModelBase):
-    """TODO"""
+    """
+    BERT4Rec model.
+
+    n_blocks: int, default 1
+        Number of transformer blocks.
+    n_heads: int, default 1
+        Number of attention heads.
+    n_factors: int, default 128
+        Latent embeddings size.
+    use_pos_emb: bool, default ``True``
+        If ``True``, adds learnable positional encoding to session item embeddings.
+    use_causal_attn: bool, default ``True``
+        If ``True``, uses causal mask as attn_mask in Multi-head Attention.
+    use_key_padding_mask: bool, default ``False``
+        If ``True``, uses key_padding_mask in Multi-head Attention.
+    dropout_rate: float, default 0.2
+        Probability of a hidden unit to be zeroed.
+    session_max_len: int, default 32
+        Maximum length of user sequence.
+    train_min_user_interactions: int, default 2
+        Minimum number of interactions user should have to be used for training. Should be greater than 1.
+    mask_prob: float, default 0.15
+        Probability of masking an item in interactions sequence.
+    dataloader_num_workers: int, default 0
+        Number of loader worker processes.
+    batch_size: int, default 128
+        How many samples per batch to load.
+    loss: str, default "softmax"
+        Loss function.
+    n_negatives: int, default 1
+        Number of negatives for BCE and gBCE losses.
+    gbce_t: float, default 0.2
+        Calibration parameter for gBCE loss.
+    lr: float, default 0.01
+        Learning rate.
+    epochs: int, default 3
+        Number of training epochs.
+    verbose: int, default 0
+        Verbosity level.
+    deterministic: bool, default ``False``
+        If ``True``, sets deterministic algorithms for PyTorch operations.
+        Use `pytorch_lightning.seed_everything` together with this parameter to fix the random state.
+    recommend_device: Union[str, Accelerator], default "auto"
+        Device for recommend. Used at predict_step of lightning module.
+        If you want to change this parameter after model is initialized,
+        you can manually assign new value to model `recommend_device` attribute.
+    recommend_n_threads: int, default 0
+        Number of threads to use in ranker.
+        If you want to change this parameter after model is initialized,
+        you can manually assign new value to model `recommend_n_threads` attribute.
+    recommend_use_gpu_ranking: bool, default ``True``
+        If ``True`` and HAS_CUDA ``True``, sets use_gpu=True in ImplicitRanker.rank.
+        If you want to change this parameter after model is initialized,
+        you can manually assign new value to model `recommend_use_gpu_ranking` attribute.
+    trainer: Optional(Trainer), default None
+        Which trainer to use for training.
+        If trainer is None, default pytorch_lightning Trainer is created.
+    item_net_block_types: Type(ItemNetBase), default (IdEmbeddingsItemNet, CatFeaturesItemNet)
+        Type of network returning item enbeddings.
+        (IdEmbeddingsItemNet,) - item embeddings based on ids.
+        (, CatFeaturesItemNet) - item embeddings based on categorical features.
+        (IdEmbeddingsItemNet, CatFeaturesItemNet) - item embeddings based on ids and categorical features.
+    pos_encoding_type: Type(PositionalEncodingBase), default `LearnableInversePositionalEncoding`
+        Type of positional encoding.
+    transformer_layers_type: Type(TransformerLayersBase), default `SasRecTransformerLayers`
+        Type of transformer layers architecture.
+    data_preparator_type: Type(SessionEncoderDataPreparatorBase), default `SasRecDataPreparator`
+        Type of data preparator used for dataset processing and dataloader creation.
+    lightning_module_type: Type(SessionEncoderLightningModuleBase), default `SessionEncoderLightningModule`
+        Type of lightning module defining training procedure.
+    """
 
     def __init__(  # pylint: disable=too-many-arguments, too-many-locals
         self,
@@ -158,7 +180,7 @@ class BERT4RecModel(TransformerModelBase):
         n_heads: int = 1,
         n_factors: int = 128,
         use_pos_emb: bool = True,
-        use_causal_attn: bool = False,
+        use_causal_attn: bool = False,  # TODO: remove?
         use_key_padding_mask: bool = True,
         dropout_rate: float = 0.2,
         epochs: int = 3,
@@ -174,12 +196,12 @@ class BERT4RecModel(TransformerModelBase):
         gbce_t: float = 0.2,
         lr: float = 0.01,
         dataloader_num_workers: int = 0,
-        train_min_user_interaction: int = 2,
+        train_min_user_interactions: int = 2,
         mask_prob: float = 0.15,
         trainer: tp.Optional[Trainer] = None,
         item_net_block_types: tp.Sequence[tp.Type[ItemNetBase]] = (IdEmbeddingsItemNet, CatFeaturesItemNet),
         pos_encoding_type: tp.Type[PositionalEncodingBase] = LearnableInversePositionalEncoding,
-        transformer_layers_type: tp.Type[TransformerLayersBase] = BERT4RecTransformerLayers,
+        transformer_layers_type: tp.Type[TransformerLayersBase] = PreLNTransformerLayers,
         data_preparator_type: tp.Type[BERT4RecDataPreparator] = BERT4RecDataPreparator,
         lightning_module_type: tp.Type[SessionEncoderLightningModuleBase] = SessionEncoderLightningModule,
     ):
@@ -213,7 +235,7 @@ class BERT4RecModel(TransformerModelBase):
             n_negatives=n_negatives if loss != "softmax" else None,
             batch_size=batch_size,
             dataloader_num_workers=dataloader_num_workers,
-            train_min_user_interactions=train_min_user_interaction,
+            train_min_user_interactions=train_min_user_interactions,
             item_extra_tokens=(PADDING_VALUE, MASKING_VALUE),
             mask_prob=mask_prob,
         )
