+++ rectools/models/nn/transformer_data_preparator.py
@@ -1,0 +1,289 @@
+import typing as tp
+import warnings
+
+import numpy as np
+import pandas as pd
+import torch
+from scipy import sparse
+from torch.utils.data import DataLoader
+from torch.utils.data import Dataset as TorchDataset
+
+from rectools import Columns, ExternalIds
+from rectools.dataset import Dataset, Interactions
+from rectools.dataset.features import SparseFeatures
+from rectools.dataset.identifiers import IdMap
+
+
+class SequenceDataset(TorchDataset):
+    """
+    Dataset for sequential data.
+
+    Parameters
+    ----------
+    sessions: List[List[int]]
+        User sessions in the form of sequences of items ids.
+    weights: List[List[float]]
+        Weight of each interaction from the session.
+    """
+
+    def __init__(self, sessions: tp.List[tp.List[int]], weights: tp.List[tp.List[float]]):
+        self.sessions = sessions
+        self.weights = weights
+
+    def __len__(self) -> int:
+        return len(self.sessions)
+
+    def __getitem__(self, index: int) -> tp.Tuple[tp.List[int], tp.List[float]]:
+        session = self.sessions[index]  # [session_len]
+        weights = self.weights[index]  # [session_len]
+        return session, weights
+
+    @classmethod
+    def from_interactions(
+        cls,
+        interactions: pd.DataFrame,
+    ) -> "SequenceDataset":
+        """
+        Group interactions by user.
+        Construct SequenceDataset from grouped interactions.
+
+        Parameters
+        ----------
+        interactions: pd.DataFrame
+            User-item interactions.
+        """
+        sessions = (
+            interactions.sort_values(Columns.Datetime)
+            .groupby(Columns.User, sort=True)[[Columns.Item, Columns.Weight]]
+            .agg(list)
+        )
+        sessions, weights = (
+            sessions[Columns.Item].to_list(),
+            sessions[Columns.Weight].to_list(),
+        )
+
+        return cls(sessions=sessions, weights=weights)
+
+
+class SessionEncoderDataPreparatorBase:
+    """
+    Base class for data preparator. To change train/recommend dataset processing, train/recommend dataloaders inherit
+    from this class and pass your custom data preparator to your model parameters.
+
+    Parameters
+    ----------
+    session_max_len: int
+        Maximum length of user sequence.
+    batch_size: int
+        How many samples per batch to load.
+    dataloader_num_workers: int
+        Number of loader worker processes.
+    item_extra_tokens: Sequence(Hashable)
+        Which element to use for sequence padding.
+    shuffle_train: bool, default True
+        If ``True``, reshuffles data at each epoch.
+    train_min_user_interactions: int, default 2
+        Minimum length of user sequence. Cannot be less than 2.
+    """
+
+    def __init__(
+        self,
+        session_max_len: int,
+        batch_size: int,
+        dataloader_num_workers: int,
+        item_extra_tokens: tp.Sequence[tp.Hashable],
+        shuffle_train: bool = True,
+        train_min_user_interactions: int = 2,
+        n_negatives: tp.Optional[int] = None,
+    ) -> None:
+        """TODO"""
+        self.item_id_map: IdMap
+        self.extra_token_ids: tp.Dict
+        self.session_max_len = session_max_len
+        self.n_negatives = n_negatives
+        self.batch_size = batch_size
+        self.dataloader_num_workers = dataloader_num_workers
+        self.train_min_user_interactions = train_min_user_interactions
+        self.item_extra_tokens = item_extra_tokens
+        self.shuffle_train = shuffle_train
+
+    def get_known_items_sorted_internal_ids(self) -> np.ndarray:
+        """Return internal item ids from processed dataset in sorted order."""
+        return self.item_id_map.get_sorted_internal()[self.n_item_extra_tokens :]
+
+    def get_known_item_ids(self) -> np.ndarray:
+        """Return external item ids from processed dataset in sorted order."""
+        return self.item_id_map.get_external_sorted_by_internal()[self.n_item_extra_tokens :]
+
+    @property
+    def n_item_extra_tokens(self) -> int:
+        """Return number of padding elements"""
+        return len(self.item_extra_tokens)
+
+    def process_dataset_train(self, dataset: Dataset) -> Dataset:
+        """Filter interactions and process features"""
+        interactions = dataset.get_raw_interactions()
+
+        # Filter interactions
+        user_stats = interactions[Columns.User].value_counts()
+        users = user_stats[user_stats >= self.train_min_user_interactions].index
+        interactions = interactions[(interactions[Columns.User].isin(users))]
+        interactions = interactions.sort_values(Columns.Datetime).groupby(Columns.User).tail(self.session_max_len + 1)
+
+        # Construct dataset
+        # TODO: user features are dropped for now
+        user_id_map = IdMap.from_values(interactions[Columns.User].values)
+        item_id_map = IdMap.from_values(self.item_extra_tokens)
+        item_id_map = item_id_map.add_ids(interactions[Columns.Item])
+
+        # get item features
+        item_features = None
+        if dataset.item_features is not None:
+            item_features = dataset.item_features
+            # TODO: remove assumption on SparseFeatures and add Dense Features support
+            if not isinstance(item_features, SparseFeatures):
+                raise ValueError("`item_features` in `dataset` must be `SparseFeatures` instance.")
+
+            internal_ids = dataset.item_id_map.convert_to_internal(
+                item_id_map.get_external_sorted_by_internal()[self.n_item_extra_tokens :]
+            )
+            sorted_item_features = item_features.take(internal_ids)
+
+            dtype = sorted_item_features.values.dtype
+            n_features = sorted_item_features.values.shape[1]
+            extra_token_feature_values = sparse.csr_matrix((self.n_item_extra_tokens, n_features), dtype=dtype)
+
+            full_feature_values: sparse.scr_matrix = sparse.vstack(
+                [extra_token_feature_values, sorted_item_features.values], format="csr"
+            )
+
+            item_features = SparseFeatures.from_iterables(values=full_feature_values, names=item_features.names)
+
+        interactions = Interactions.from_raw(interactions, user_id_map, item_id_map)
+
+        dataset = Dataset(user_id_map, item_id_map, interactions, item_features=item_features)
+
+        self.item_id_map = dataset.item_id_map
+
+        extra_token_ids = self.item_id_map.convert_to_internal(self.item_extra_tokens)
+        self.extra_token_ids = dict(zip(self.item_extra_tokens, extra_token_ids))
+        return dataset
+
+    def get_dataloader_train(self, processed_dataset: Dataset) -> DataLoader:
+        """
+        Construct train dataloader from processed dataset.
+
+        Parameters
+        ----------
+        processed_dataset: Dataset
+            RecTools dataset prepared for training.
+
+        Returns
+        -------
+        DataLoader
+            Train dataloader.
+        """
+        sequence_dataset = SequenceDataset.from_interactions(processed_dataset.interactions.df)
+        train_dataloader = DataLoader(
+            sequence_dataset,
+            collate_fn=self._collate_fn_train,
+            batch_size=self.batch_size,
+            num_workers=self.dataloader_num_workers,
+            shuffle=self.shuffle_train,
+        )
+        return train_dataloader
+
+    def get_dataloader_recommend(self, dataset: Dataset) -> DataLoader:
+        """TODO"""
+        sequence_dataset = SequenceDataset.from_interactions(dataset.interactions.df)
+        recommend_dataloader = DataLoader(
+            sequence_dataset,
+            batch_size=self.batch_size,
+            collate_fn=self._collate_fn_recommend,
+            num_workers=self.dataloader_num_workers,
+            shuffle=False,
+        )
+        return recommend_dataloader
+
+    def transform_dataset_u2i(self, dataset: Dataset, users: ExternalIds) -> Dataset:
+        """
+        Process dataset for u2i recommendations.
+        Filter out interactions and adapt id maps.
+        All users beyond target users for recommendations are dropped.
+        All target users that do not have at least one known item in interactions are dropped.
+
+        Parameters
+        ----------
+        dataset: Dataset
+            RecTools dataset.
+        users: ExternalIds
+            Array of external user ids to recommend for.
+
+        Returns
+        -------
+        Dataset
+            Processed RecTools dataset.
+            Final dataset will consist only of model known items during fit and only of required
+            (and supported) target users for recommendations.
+            Final user_id_map is an enumerated list of supported (filtered) target users.
+            Final item_id_map is model item_id_map constructed during training.
+        """
+        # Filter interactions in dataset internal ids
+        interactions = dataset.interactions.df
+        users_internal = dataset.user_id_map.convert_to_internal(users, strict=False)
+        items_internal = dataset.item_id_map.convert_to_internal(self.get_known_item_ids(), strict=False)
+        interactions = interactions[interactions[Columns.User].isin(users_internal)]  # todo: fast_isin
+        interactions = interactions[interactions[Columns.Item].isin(items_internal)]
+
+        # Convert to external ids
+        interactions[Columns.Item] = dataset.item_id_map.convert_to_external(interactions[Columns.Item])
+        interactions[Columns.User] = dataset.user_id_map.convert_to_external(interactions[Columns.User])
+
+        # Prepare new user id mapping
+        rec_user_id_map = IdMap.from_values(interactions[Columns.User])
+
+        # Construct dataset
+        # TODO: For now features are dropped because model doesn't support them
+        n_filtered = len(users) - rec_user_id_map.size
+        if n_filtered > 0:
+            explanation = f"""{n_filtered} target users were considered cold because of missing known items"""
+            warnings.warn(explanation)
+        filtered_interactions = Interactions.from_raw(interactions, rec_user_id_map, self.item_id_map)
+        filtered_dataset = Dataset(rec_user_id_map, self.item_id_map, filtered_interactions)
+        return filtered_dataset
+
+    def transform_dataset_i2i(self, dataset: Dataset) -> Dataset:
+        """
+        Process dataset for i2i recommendations.
+        Filter out interactions and adapt id maps.
+
+        Parameters
+        ----------
+        dataset: Dataset
+            RecTools dataset.
+
+        Returns
+        -------
+        Dataset
+            Processed RecTools dataset.
+            Final dataset will consist only of model known items during fit.
+            Final user_id_map is the same as dataset original.
+            Final item_id_map is model item_id_map constructed during training.
+        """
+        interactions = dataset.get_raw_interactions()
+        interactions = interactions[interactions[Columns.Item].isin(self.get_known_item_ids())]
+        filtered_interactions = Interactions.from_raw(interactions, dataset.user_id_map, self.item_id_map)
+        filtered_dataset = Dataset(dataset.user_id_map, self.item_id_map, filtered_interactions)
+        return filtered_dataset
+
+    def _collate_fn_train(
+        self,
+        batch: tp.List[tp.Tuple[tp.List[int], tp.List[float]]],
+    ) -> tp.Dict[str, torch.Tensor]:
+        raise NotImplementedError()
+
+    def _collate_fn_recommend(
+        self,
+        batch: tp.List[tp.Tuple[tp.List[int], tp.List[float]]],
+    ) -> tp.Dict[str, torch.Tensor]:
+        raise NotImplementedError()
